requirements.txt
pandas
numpy
faiss-cpu
streamlit
sentence-transformers
langchain
langchain-community
python-dotenv
google-generativeai
